pyproject.toml
[tool.poetry]
name = "hrr-analysis"
version = "1.3.2"
description = "Personalized Freediving Recovery Analysis & ML Feasibility Study"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
numpy = ">=1.21.0,<2.0.0"
pandas = ">=1.3.0,<3.0.0"
scikit-learn = ">=1.0.0,<2.0.0"
xgboost = ">=1.5.0"
matplotlib = ">=3.5.0"
seaborn = ">=0.11.0"
shap = ">=0.40.0"
joblib = ">=1.1.0"
tqdm = ">=4.62.0"
pyarrow = "^15.0.0" # For parquet file handling

# MLOps Tooling
dvc = {extras = ["gs"], version = "^3.0"}
mlflow = "^2.9.0"
pyyaml = "^6.0" # For reading config files
lime = "^0.2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
flake8 = "^6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"